requirements.txt
flask-2.2.2
requests-2.28.1


